+++ realm/realm-library/src/androidTest/java/io/realm/RealmTests.java
@@ -61,7 +61,6 @@ import java.util.concurrent.atomic.AtomicReference;
 import io.realm.entities.AllJavaTypes;
 import io.realm.entities.AllTypes;
 import io.realm.entities.AllTypesPrimaryKey;
-import io.realm.entities.AnnotationIndexTypes;
 import io.realm.entities.Cat;
 import io.realm.entities.CyclicType;
 import io.realm.entities.CyclicTypePrimaryKey;
@@ -514,117 +513,6 @@ public class RealmTests {
     }
 
     @Test
-    public void allObjects() {
-        populateTestRealm();
-        RealmResults<AllTypes> resultList = realm.allObjects(AllTypes.class);
-        assertEquals("Realm.get is returning wrong result set", TEST_DATA_SIZE, resultList.size());
-    }
-
-    @Test
-    public void allObjectsSorted() {
-        populateTestRealm();
-        RealmResults<AllTypes> sortedList = realm.allObjectsSorted(AllTypes.class, AllTypes.FIELD_STRING, Sort.ASCENDING);
-        assertEquals(TEST_DATA_SIZE, sortedList.size());
-        assertEquals("test data 0", sortedList.first().getColumnString());
-
-        RealmResults<AllTypes> reverseList = realm.allObjectsSorted(AllTypes.class, AllTypes.FIELD_STRING, Sort.DESCENDING);
-        assertEquals(TEST_DATA_SIZE, reverseList.size());
-        assertEquals("test data 0", reverseList.last().getColumnString());
-
-        try {
-            realm.allObjectsSorted(AllTypes.class, "invalid", Sort.ASCENDING);
-            fail();
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
-    @Test
-    public void allObjectsSorted_singleField() {
-        populateTestRealm();
-        RealmResults<AllTypes> sortedList = realm.allObjectsSorted(AllTypes.class,
-                new String[]{AllTypes.FIELD_LONG},
-                new Sort[]{Sort.DESCENDING});
-        assertEquals(TEST_DATA_SIZE, sortedList.size());
-        assertEquals(TEST_DATA_SIZE - 1, sortedList.first().getColumnLong());
-        assertEquals(0, sortedList.last().getColumnLong());
-    }
-
-    @Test
-    public void allObjectsSorted_twoFields() {
-        TestHelper.populateForMultiSort(realm);
-
-        RealmResults<AllTypes> results1 = realm.allObjectsSorted(AllTypes.class,
-                new String[]{AllTypes.FIELD_STRING, AllTypes.FIELD_LONG},
-                new Sort[]{Sort.ASCENDING, Sort.ASCENDING});
-
-        assertEquals(3, results1.size());
-
-        assertEquals("Adam", results1.get(0).getColumnString());
-        assertEquals(4, results1.get(0).getColumnLong());
-
-        assertEquals("Adam", results1.get(1).getColumnString());
-        assertEquals(5, results1.get(1).getColumnLong());
-
-        assertEquals("Brian", results1.get(2).getColumnString());
-        assertEquals(4, results1.get(2).getColumnLong());
-
-        RealmResults<AllTypes> results2 = realm.allObjectsSorted(AllTypes.class,
-                new String[]{AllTypes.FIELD_LONG, AllTypes.FIELD_STRING},
-                new Sort[]{Sort.ASCENDING, Sort.ASCENDING});
-
-        assertEquals(3, results2.size());
-
-        assertEquals("Adam", results2.get(0).getColumnString());
-        assertEquals(4, results2.get(0).getColumnLong());
-
-        assertEquals("Brian", results2.get(1).getColumnString());
-        assertEquals(4, results2.get(1).getColumnLong());
-
-        assertEquals("Adam", results2.get(2).getColumnString());
-        assertEquals(5, results2.get(2).getColumnLong());
-    }
-
-    @Test
-    public void allObjectsSorted_failures() {
-        // zero fields specified
-        try {
-            realm.allObjectsSorted(AllTypes.class, new String[]{}, new Sort[]{});
-            fail();
-        } catch (IllegalArgumentException ignored) {
-        }
-
-        // number of fields and sorting orders don't match
-        try {
-            realm.allObjectsSorted(AllTypes.class,
-                    new String[]{AllTypes.FIELD_STRING},
-                    new Sort[]{Sort.ASCENDING, Sort.ASCENDING});
-            fail();
-        } catch (IllegalArgumentException ignored) {
-        }
-
-        // null is not allowed
-        try {
-            realm.allObjectsSorted(AllTypes.class, null, (Sort[]) null);
-            fail();
-        } catch (IllegalArgumentException ignored) {
-        }
-        try {
-            realm.allObjectsSorted(AllTypes.class, new String[]{AllTypes.FIELD_STRING}, null);
-            fail();
-        } catch (IllegalArgumentException ignored) {
-        }
-
-        // non-existing field name
-        try {
-            realm.allObjectsSorted(AllTypes.class,
-                    new String[]{AllTypes.FIELD_STRING, "dont-exist"},
-                    new Sort[]{Sort.ASCENDING, Sort.ASCENDING});
-            fail();
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
-    @Test
     public void beginTransaction() throws IOException {
         populateTestRealm();
 
@@ -661,7 +549,6 @@ public class RealmTests {
         METHOD_CANCEL,
         METHOD_DELETE_TYPE,
         METHOD_DELETE_ALL,
-        METHOD_DISTINCT,
         METHOD_CREATE_OBJECT,
         METHOD_COPY_TO_REALM,
         METHOD_COPY_TO_REALM_OR_UPDATE,
@@ -697,9 +584,6 @@ public class RealmTests {
                             break;
                         case METHOD_DELETE_ALL:
                             realm.deleteAll();
-                            break;
-                        case METHOD_DISTINCT:
-                            realm.distinct(AllTypesPrimaryKey.class, "columnLong");
                             break;
                         case METHOD_CREATE_OBJECT:
                             realm.createObject(AllTypes.class);
@@ -781,7 +665,7 @@ public class RealmTests {
         realm.beginTransaction();
         realm.createObject(AllTypes.class);
         realm.cancelTransaction();
-        assertEquals(TEST_DATA_SIZE, realm.allObjects(AllTypes.class).size());
+        assertEquals(TEST_DATA_SIZE, realm.where(AllTypes.class).count());
 
         try {
             realm.cancelTransaction();
@@ -803,7 +687,7 @@ public class RealmTests {
 
     @Test
     public void executeTransaction_success() {
-        assertEquals(0, realm.allObjects(Owner.class).size());
+        assertEquals(0, realm.where(Owner.class).count());
         realm.executeTransaction(new Realm.Transaction() {
             @Override
             public void execute(Realm realm) {
@@ -811,14 +695,14 @@ public class RealmTests {
                 owner.setName("Owner");
             }
         });
-        assertEquals(1, realm.allObjects(Owner.class).size());
+        assertEquals(1, realm.where(Owner.class).count());
     }
 
     @Test
     public void executeTransaction_canceled() {
         final AtomicReference<RuntimeException> thrownException = new AtomicReference<>(null);
 
-        assertEquals(0, realm.allObjects(Owner.class).size());
+        assertEquals(0, realm.where(Owner.class).count());
         try {
             realm.executeTransaction(new Realm.Transaction() {
                 @Override
@@ -833,12 +717,12 @@ public class RealmTests {
             //noinspection ThrowableResultOfMethodCallIgnored
             assertTrue(e == thrownException.get());
         }
-        assertEquals(0, realm.allObjects(Owner.class).size());
+        assertEquals(0, realm.where(Owner.class).count());
     }
 
     @Test
     public void executeTransaction_cancelInsideClosureThrowsException() {
-        assertEquals(0, realm.allObjects(Owner.class).size());
+        assertEquals(0, realm.where(Owner.class).count());
         TestHelper.TestLogger testLogger = new TestHelper.TestLogger();
         try {
             RealmLog.add(testLogger);
@@ -857,7 +741,7 @@ public class RealmTests {
         } finally {
             RealmLog.remove(testLogger);
         }
-        assertEquals(0, realm.allObjects(Owner.class).size());
+        assertEquals(0, realm.where(Owner.class).count());
     }
 
     @Test
@@ -906,7 +790,7 @@ public class RealmTests {
         assertTrue(language, file.exists());
 
         Realm realm2 = Realm.getInstance(realmConfig);
-        Dog dog2 = realm2.allObjects(Dog.class).first();
+        Dog dog2 = realm2.where(Dog.class).findFirst();
         assertEquals(language, "Rex", dog2.getName());
         realm2.close();
     }
@@ -1006,7 +890,7 @@ public class RealmTests {
             stringOnly.setChars(test_char);
             realm.commitTransaction();
 
-            realm.allObjects(StringOnly.class).get(0).getChars();
+            realm.where(StringOnly.class).findFirst().getChars();
 
             realm.beginTransaction();
             realm.delete(StringOnly.class);
@@ -1074,7 +958,7 @@ public class RealmTests {
         realm = Realm.getInstance(configFactory.createConfiguration());
         realm.beginTransaction();
         AllTypes allTypes = realm.createObject(AllTypes.class);
-        RealmResults<AllTypes> queryResult = realm.allObjects(AllTypes.class);
+        RealmResults<AllTypes> queryResult = realm.where(AllTypes.class).findAll();
         assertEquals(allTypes, queryResult.get(0));
         realm.commitTransaction();
         realm.close(); // This might not close the Realm if the reference count is wrong
@@ -1115,7 +999,7 @@ public class RealmTests {
         try {
             // Contents is copied too
             realm2 = Realm.getInstance(configB);
-            RealmResults<AllTypes> results = realm2.allObjects(AllTypes.class);
+            RealmResults<AllTypes> results = realm2.where(AllTypes.class).findAll();
             assertEquals(1, results.size());
             assertEquals("Hello World", results.first().getColumnString());
         } finally {
@@ -1288,7 +1172,7 @@ public class RealmTests {
 
         assertEquals("One", realmObject.getName());
         assertEquals("Two", realmObject.getObject().getName());
-        assertEquals(2, realm.allObjects(CyclicType.class).size());
+        assertEquals(2, realm.where(CyclicType.class).count());
     }
 
     @Test
@@ -1305,7 +1189,7 @@ public class RealmTests {
         realm.commitTransaction();
 
         assertEquals("One", realmObject.getName());
-        assertEquals(2, realm.allObjects(CyclicType.class).size());
+        assertEquals(2, realm.where(CyclicType.class).count());
     }
 
     // Check that if a field has a null value it gets converted to the default value for that type
@@ -1598,7 +1482,7 @@ public class RealmTests {
             }
         });
 
-        assertEquals(2, realm.allObjects(PrimaryKeyAsLong.class).size());
+        assertEquals(2, realm.where(PrimaryKeyAsLong.class).count());
     }
 
     @Test
@@ -1634,8 +1518,8 @@ public class RealmTests {
             }
         });
 
-        assertEquals(1, realm.allObjects(AllTypesPrimaryKey.class).size());
-        AllTypesPrimaryKey obj = realm.allObjects(AllTypesPrimaryKey.class).first();
+        assertEquals(1, realm.where(AllTypesPrimaryKey.class).count());
+        AllTypesPrimaryKey obj = realm.where(AllTypesPrimaryKey.class).findFirst();
 
         // Check that the the only element has all its properties updated
         assertEquals("Bar", obj.getColumnString());
@@ -1670,7 +1554,7 @@ public class RealmTests {
         realm.copyToRealmOrUpdate(oneCyclicType);
         realm.commitTransaction();
 
-        assertEquals(2, realm.allObjects(CyclicTypePrimaryKey.class).size());
+        assertEquals(2, realm.where(CyclicTypePrimaryKey.class).count());
         assertEquals("Three", realm.where(CyclicTypePrimaryKey.class).equalTo("id", 1).findFirst().getName());
     }
 
@@ -1699,9 +1583,9 @@ public class RealmTests {
             }
         });
 
-        assertEquals(1, realm.allObjects(AllTypesPrimaryKey.class).size());
+        assertEquals(1, realm.where(AllTypesPrimaryKey.class).count());
 
-        AllTypesPrimaryKey obj = realm.allObjects(AllTypesPrimaryKey.class).first();
+        AllTypesPrimaryKey obj = realm.where(AllTypesPrimaryKey.class).findFirst();
         assertNull(obj.getColumnString());
         assertEquals(1, obj.getColumnLong());
         assertEquals(0.0F, obj.getColumnFloat(), 0);
@@ -1734,8 +1618,8 @@ public class RealmTests {
             }
         });
 
-        assertEquals(1, realm.allObjects(AllTypesPrimaryKey.class).size());
-        assertEquals(4, realm.allObjects(DogPrimaryKey.class).size());
+        assertEquals(1, realm.where(AllTypesPrimaryKey.class).count());
+        assertEquals(4, realm.where(DogPrimaryKey.class).count());
     }
 
     @Test
@@ -1789,8 +1673,8 @@ public class RealmTests {
             }
         });
 
-        assertEquals(1, realm.allObjects(PrimaryKeyAsLong.class).size());
-        assertEquals("Baz", realm.allObjects(PrimaryKeyAsLong.class).first().getName());
+        assertEquals(1, realm.where(PrimaryKeyAsLong.class).count());
+        assertEquals("Baz", realm.where(PrimaryKeyAsLong.class).findFirst().getName());
     }
 
     // Tests that a collection of objects with references all gets copied.
@@ -1810,8 +1694,8 @@ public class RealmTests {
         realm.copyToRealmOrUpdate(Arrays.asList(allTypes1, allTypes2));
         realm.commitTransaction();
 
-        assertEquals(2, realm.allObjects(AllTypesPrimaryKey.class).size());
-        assertEquals(1, realm.allObjects(DogPrimaryKey.class).size());
+        assertEquals(2, realm.where(AllTypesPrimaryKey.class).count());
+        assertEquals(1, realm.where(DogPrimaryKey.class).count());
     }
 
     @Test
@@ -2578,253 +2462,6 @@ public class RealmTests {
         assertTrue(typeListenerWasCalled.get());
     }
 
-    private void populateForDistinct(Realm realm, long numberOfBlocks, long numberOfObjects, boolean withNull) {
-        realm.beginTransaction();
-        for (int i = 0; i < numberOfObjects * numberOfBlocks; i++) {
-            for (int j = 0; j < numberOfBlocks; j++) {
-                AnnotationIndexTypes obj = realm.createObject(AnnotationIndexTypes.class);
-                obj.setIndexBoolean(j % 2 == 0);
-                obj.setIndexLong(j);
-                obj.setIndexDate(withNull ? null : new Date(1000 * (long) j));
-                obj.setIndexString(withNull ? null : "Test " + j);
-                obj.setNotIndexBoolean(j % 2 == 0);
-                obj.setNotIndexLong(j);
-                obj.setNotIndexDate(withNull ? null : new Date(1000 * (long) j));
-                obj.setNotIndexString(withNull ? null : "Test " + j);
-            }
-        }
-        realm.commitTransaction();
-    }
-
-    private void populateForDistinctInvalidTypesLinked(Realm realm) {
-        realm.beginTransaction();
-        AllJavaTypes notEmpty = new AllJavaTypes();
-        notEmpty.setFieldBinary(new byte[]{1, 2, 3});
-        notEmpty.setFieldObject(notEmpty);
-        notEmpty.setFieldList(new RealmList<AllJavaTypes>(notEmpty));
-        realm.copyToRealm(notEmpty);
-        realm.commitTransaction();
-    }
-
-    // Realm.distinct(): requires indexing, and type = boolean, integer, date, string
-    @Test
-    public void distinct() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10; // must be greater than 1
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, false);
-
-        RealmResults<AnnotationIndexTypes> distinctBool = realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_BOOL);
-        assertEquals(2, distinctBool.size());
-        for (String field : new String[]{AnnotationIndexTypes.FIELD_INDEX_LONG, AnnotationIndexTypes.FIELD_INDEX_DATE, AnnotationIndexTypes.FIELD_INDEX_STRING}) {
-            RealmResults<AnnotationIndexTypes> distinct = realm.distinct(AnnotationIndexTypes.class, field);
-            assertEquals(field, numberOfBlocks, distinct.size());
-        }
-    }
-
-    @Test
-    public void distinct_withNullValues() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10; // must be greater than 1
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, true);
-
-        for (String field : new String[]{AnnotationIndexTypes.FIELD_INDEX_DATE, AnnotationIndexTypes.FIELD_INDEX_STRING}) {
-            RealmResults<AnnotationIndexTypes> distinct = realm.distinct(AnnotationIndexTypes.class, field);
-            assertEquals(field, 1, distinct.size());
-        }
-    }
-
-    @Test
-    public void distinct_notIndexedFieldsThrows() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10; // must be greater than 1
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, false);
-
-        for (String field : AnnotationIndexTypes.NOT_INDEX_FIELDS) {
-            try {
-                realm.distinct(AnnotationIndexTypes.class, field);
-                fail(field);
-            } catch (IllegalArgumentException ignored) {
-            }
-        }
-    }
-
-    @Test
-    public void distinct_unknownFieldThrows() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10; // must be greater than 1
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, false);
-        thrown.expect(IllegalArgumentException.class);
-
-        realm.distinct(AnnotationIndexTypes.class, "doesNotExist");
-    }
-
-    @Test
-    public void distinct_invalidTypeThrows() {
-        populateTestRealm();
-
-        for (String field : new String[]{AllTypes.FIELD_REALMOBJECT, AllTypes.FIELD_REALMLIST, AllTypes.FIELD_DOUBLE, AllTypes.FIELD_FLOAT}) {
-            try {
-                realm.distinct(AllTypes.class, field);
-                fail(field);
-            } catch (IllegalArgumentException ignored) {
-            }
-        }
-    }
-
-    @Test
-    public void distinctMultiArgs() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10; // must be greater than 1
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, false);
-
-        RealmResults<AnnotationIndexTypes> distinctMulti = realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_BOOL, AnnotationIndexTypes.INDEX_FIELDS);
-        assertEquals(numberOfBlocks, distinctMulti.size());
-    }
-
-    @Test
-    public void distinctMultiArgs_switchedFieldsOrder() {
-        final long numberOfBlocks = 25;
-        TestHelper.populateForDistinctFieldsOrder(realm, numberOfBlocks);
-
-        // Regardless of the block size defined above, the output size is expected to be the same, 4 in this case, due to receiving unique combinations of tuples
-        RealmResults<AnnotationIndexTypes> distinctStringLong = realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_STRING, AnnotationIndexTypes.FIELD_INDEX_LONG);
-        RealmResults<AnnotationIndexTypes> distinctLongString = realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_LONG, AnnotationIndexTypes.FIELD_INDEX_STRING);
-        assertEquals(4, distinctStringLong.size());
-        assertEquals(4, distinctLongString.size());
-        assertEquals(distinctStringLong.size(), distinctLongString.size());
-    }
-
-    @Test
-    public void distinctMultiArgs_emptyFields() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10;
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, false);
-
-        // an empty string field in the middle
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_BOOL, "", AnnotationIndexTypes.FIELD_INDEX_INT);
-        } catch (IllegalArgumentException ignored) {
-        }
-        // an empty string field at the end
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_BOOL, AnnotationIndexTypes.FIELD_INDEX_INT, "");
-        } catch (IllegalArgumentException ignored) {
-        }
-        // a null string field in the middle
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_BOOL, (String)null, AnnotationIndexTypes.FIELD_INDEX_INT);
-        } catch (IllegalArgumentException ignored) {
-        }
-        // a null string field at the end
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_BOOL, AnnotationIndexTypes.FIELD_INDEX_INT, (String)null);
-        } catch (IllegalArgumentException ignored) {
-        }
-        // (String)null makes varargs a null array.
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_BOOL, (String)null);
-        } catch (IllegalArgumentException ignored) {
-        }
-        // Two (String)null for first and varargs fields
-        try {
-            realm.distinct(AnnotationIndexTypes.class, (String)null, (String)null);
-        } catch (IllegalArgumentException ignored) {
-        }
-        // "" & (String)null combination
-        try {
-            realm.distinct(AnnotationIndexTypes.class, "", (String)null);
-        } catch (IllegalArgumentException ignored) {
-        }
-        // "" & (String)null combination
-        try {
-            realm.distinct(AnnotationIndexTypes.class, (String)null, "");
-        } catch (IllegalArgumentException ignored) {
-        }
-        // Two empty fields tests
-        try {
-            realm.distinct(AnnotationIndexTypes.class, "", "");
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
-    @Test
-    public void distinctMultiArgs_withNullValues() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10;
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, true);
-
-        RealmResults<AnnotationIndexTypes> distinctMulti = realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_DATE, AnnotationIndexTypes.FIELD_INDEX_STRING);
-        assertEquals(1, distinctMulti.size());
-    }
-
-    @Test
-    public void distinctMultiArgs_notIndexedFields() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10;
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, false);
-
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_NOT_INDEX_STRING, AnnotationIndexTypes.NOT_INDEX_FIELDS);
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
-    @Test
-    public void distinctMultiArgs_doesNotExistField() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10;
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, false);
-
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.FIELD_INDEX_INT, AnnotationIndexTypes.NONEXISTANT_MIX_FIELDS);
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
-    @Test
-    public void distinctMultiArgs_invalidTypesFields() {
-        populateTestRealm();
-
-        try {
-            realm.distinct(AllTypes.class, AllTypes.FIELD_REALMOBJECT, AllTypes.INVALID_TYPES_FIELDS_FOR_DISTINCT);
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
-    @Test
-    public void distinctMultiArgs_indexedLinkedFields() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10;
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, true);
-
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.INDEX_LINKED_FIELD_STRING, AnnotationIndexTypes.INDEX_LINKED_FIELDS);
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
-    @Test
-    public void distinctMultiArgs_notIndexedLinkedFields() {
-        final long numberOfBlocks = 25;
-        final long numberOfObjects = 10;
-        populateForDistinct(realm, numberOfBlocks, numberOfObjects, true);
-
-        try {
-            realm.distinct(AnnotationIndexTypes.class, AnnotationIndexTypes.NOT_INDEX_LINKED_FILED_STRING, AnnotationIndexTypes.NOT_INDEX_LINKED_FIELDS);
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
-    @Test
-    public void distinctMultiArgs_invalidTypesLinkedFields() {
-        populateForDistinctInvalidTypesLinked(realm);
-
-        try {
-            realm.distinct(AllJavaTypes.class, AllJavaTypes.INVALID_LINKED_BINARY_FIELD_FOR_DISTINCT, AllJavaTypes.INVALID_LINKED_TYPES_FIELDS_FOR_DISTINCT);
-        } catch (IllegalArgumentException ignored) {
-        }
-    }
-
     @Test
     public void isInTransaction() {
         assertFalse(realm.isInTransaction());
@@ -3137,7 +2774,7 @@ public class RealmTests {
 
     @Test
     public void copyFromRealm_list_invalidDepthThrows() {
-        RealmResults<AllTypes> results = realm.allObjects(AllTypes.class);
+        RealmResults<AllTypes> results = realm.where(AllTypes.class).findAll();
         thrown.expect(IllegalArgumentException.class);
         realm.copyFromRealm(results, -1);
     }