+++ realm/realm-library/src/main/java/io/realm/DynamicRealm.java
@@ -167,28 +167,17 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Gets all objects of a specific class name.
-     *
-     * @param className the Class to get objects of.
-     * @return a {@link RealmResults} list containing the objects. If no results where found, an empty list
-     * will be returned.
-     * @see io.realm.RealmResults
+     * DEPRECATED: Use {@code dynamicRealm.where(className).findAll()} instead.
      */
+    @Deprecated
     public RealmResults<DynamicRealmObject> allObjects(String className) {
         return where(className).findAll();
     }
 
     /**
-     * Gets all objects of a specific class name sorted by a field. If no objects exist, the returned
-     * {@link RealmResults} will not be {@code null}. Use {@link RealmResults#size()} to check the number of objects
-     * instead.
-     *
-     * @param className the class to get all objects from.
-     * @param fieldName the field name to sort by.
-     * @param sortOrder how to sort the results.
-     * @return a sorted {@link RealmResults} containing the objects.
-     * @throws java.lang.IllegalArgumentException if field name does not exist.
+     * DEPRECATED: Use {@code dynamicRealm.where(className).findAll(fieldName, sortOrder)} instead.
      */
+    @Deprecated
     public RealmResults<DynamicRealmObject> allObjectsSorted(String className, String fieldName, Sort sortOrder) {
         checkIfValid();
         Table table = schema.getTable(className);
@@ -203,19 +192,9 @@ public final class DynamicRealm extends BaseRealm {
 
 
     /**
-     * Gets all objects of a specific class name sorted by two specific field names.  If no objects exist,
-     * the returned {@link RealmResults} will not be {@code null}. Use {@link RealmResults#size()} to check the number
-     * of objects instead.
-     *
-     * @param className the class to get all objects from.
-     * @param fieldName1 the first field name to sort by.
-     * @param sortOrder1 how to sort the first field.
-     * @param fieldName2 the second field name to sort by.
-     * @param sortOrder2 how to sort the second field.
-     * @return a sorted {@link RealmResults} containing the objects. If no results where found an empty list
-     * is returned.
-     * @throws java.lang.IllegalArgumentException if a field name used for sorting does not exist.
+     * DEPRECATED: Use {@code dynamicRealm.where(className).findAll(fieldName1, sortOrder1, fieldName2, sortOrder2)} instead.
      */
+    @Deprecated
     public RealmResults<DynamicRealmObject> allObjectsSorted(String className, String fieldName1,
                                                                     Sort sortOrder1, String fieldName2,
                                                                     Sort sortOrder2) {
@@ -224,17 +203,9 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Gets all objects of a specific class name sorted by multiple fields.  If no objects exist, the
-     * returned {@link RealmResults} will not be {@code null}. Use {@link RealmResults#size()} to check the number of
-     * objects instead.
-     *
-     * @param className the class to get all objects from.
-     * @param sortOrders sort ascending if Sort.ASCENDING, sort descending if Sort.DESCENDING.
-     * @param fieldNames an array of field names to sort objects by.
-     *        The objects are first sorted by fieldNames[0], then by fieldNames[1] and so forth.
-     * @return A sorted {@link RealmResults} containing the objects.
-     * @throws java.lang.IllegalArgumentException if a field name does not exist.
+     * DEPRECATED: Use {@code dynamicRealm.where(className).findAll(fieldNames[], sortOrders[])} instead.
      */
+    @Deprecated
     @SuppressWarnings("unchecked")
     public RealmResults<DynamicRealmObject> allObjectsSorted(String className, String fieldNames[], Sort sortOrders[]) {
         checkAllObjectsSortedParameters(fieldNames, sortOrders);
@@ -255,15 +226,9 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Returns a distinct set of objects of a specific class. As a Realm is unordered, it is undefined which objects are
-     * returned in case of multiple occurrences.
-     *
-     * @param className the Class to get objects of.
-     * @param fieldName the field name.
-     * @return A non-null {@link RealmResults} containing the distinct objects.
-     * @throws IllegalArgumentException if a field is null, does not exist, is an unsupported type,
-     * is not indexed, or points to linked fields.
+     * DEPRECATED: Use {@code dynamicRealm.where(className).distinct(fieldName)} instead.
      */
+    @Deprecated
     public RealmResults<DynamicRealmObject> distinct(String className, String fieldName) {
         checkIfValid();
         Table table = schema.getTable(className);
@@ -273,36 +238,18 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Returns a distinct set of objects of a specific class. As a Realm is unordered, it is undefined which objects are
-     * returned in case of multiple occurrences.
-     * This method is only available from a Looper thread.
-     *
-     * @param className the Class to get objects of.
-     * @param fieldName the field name.
-     * @return immediately a {@link RealmResults}. Users need to register a listener
-     * {@link io.realm.RealmResults#addChangeListener(RealmChangeListener)} to be notified when the
-     * query completes.
-     * @throws IllegalArgumentException if a field is null, does not exist, is an unsupported type,
-     * is not indexed, or points to linked fields.
+     * DEPRECATED: Use {@code dynamicRealm.where(className).distinctAsync(fieldName)} instead.
      */
+    @Deprecated
     public RealmResults<DynamicRealmObject> distinctAsync(String className, String fieldName) {
         checkIfValid();
         return where(className).distinctAsync(fieldName);
     }
 
     /**
-     * Returns a distinct set of objects from a specific class. When multiple distinct fields are
-     * given, all unique combinations of values in the fields will be returned. In case of multiple
-     * matches, it is undefined which object is returned. Unless the result is sorted, then the
-     * first object will be returned.
-     *
-     * @param className the Class to get objects of.
-     * @param firstFieldName first field name to use when finding distinct objects.
-     * @param remainingFieldNames remaining field names when determining all unique combinations of field values.
-     * @return a non-null {@link RealmResults} containing the distinct objects.
-     * @throws IllegalArgumentException if field names is empty or {@code null}, does not exist,
-     * is an unsupported type, or points to a linked field.
+     * DEPRECATED: Use {@code dynamicRealm.where(className).distinct(firstFieldName, remainingFieldNames)} instead.
      */
+    @Deprecated
     public RealmResults<DynamicRealmObject> distinct(String className, String firstFieldName, String... remainingFieldNames) {
         checkIfValid();
         return where(className).distinct(firstFieldName, remainingFieldNames);